\name{mandelExplorer-package}
\alias{mandelExplorer-package}
\alias{mandelExplorer}
\docType{package}
\title{
\packageTitle{mandelExplorer}
}
\description{
\packageDescription{mandelExplorer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mandelExplorer}
\packageIndices{mandelExplorer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mandelExplorer}

Maintainer: \packageMaintainer{mandelExplorer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
